requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery
redis
pydantic-settings
alembic
pydantic[email]
python-multipart
python-dotenv
pydantic-ai
openai
termcolor
PyPDF2
python-docx
httpx
textblob
scikit-learn
numpy

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0